finished video-57

+++ app/controllers/author/detail.js
@@ -1,12 +1,17 @@
 import Controller from '@ember/controller';
+import {
+  inject as service
+} from '@ember/service';
 
 export default Controller.extend({
-    actions: {
-        deleteAuthor(author) {
-            author.destroyRecord().then(() => {
-                this.transitionToRoute('author.index');
-            });
-            
-        }
+  can: service('can'),
+  actions: {
+    deleteAuthor(author) {
+      if (this.get('can').can('delete author', author)) {
+        author.destroyRecord().then(() => {
+          this.transitionToRoute('author.index');
+        });
+      }
     }
+  }
 });
